+++ src/main/java/jp/skypencil/errorprone/slf4j/PlaceholderMismatch.java
@@ -1,6 +1,10 @@
 package jp.skypencil.errorprone.slf4j;
 
 import static com.google.errorprone.BugPattern.SeverityLevel.ERROR;
+import static com.google.errorprone.matchers.Matchers.isSubtypeOf;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import com.google.auto.service.AutoService;
 import com.google.common.collect.ImmutableSet;
@@ -10,13 +14,11 @@ import com.google.errorprone.bugpatterns.BugChecker;
 import com.google.errorprone.bugpatterns.BugChecker.MethodInvocationTreeMatcher;
 import com.google.errorprone.matchers.Description;
 import com.google.errorprone.util.ASTHelpers;
+import com.sun.source.tree.ExpressionTree;
 import com.sun.source.tree.MethodInvocationTree;
 import com.sun.tools.javac.code.Symbol;
-import com.sun.tools.javac.code.Symbol.MethodSymbol;
 import com.sun.tools.javac.code.Symbol.VarSymbol;
 import com.sun.tools.javac.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 @BugPattern(
     name = "Slf4jPlaceholderMismatch",
@@ -34,8 +36,9 @@ public class PlaceholderMismatch extends BugChecker implements MethodInvocationT
 
   @Override
   public Description matchMethodInvocation(MethodInvocationTree tree, VisitorState state) {
-    MethodSymbol method = ASTHelpers.getSymbol(tree);
-    String methodName = method.toString();
+    Symbol method = ASTHelpers.getSymbol(tree.getMethodSelect());
+    String methodStr = method.toString();
+    String methodName = methodStr.substring(0, methodStr.indexOf('('));
     if (!TARGET_METHOD_NAMES.contains(methodName)) {
       return Description.NO_MATCH;
     }
@@ -45,25 +48,26 @@ public class PlaceholderMismatch extends BugChecker implements MethodInvocationT
       return Description.NO_MATCH;
     }
 
-    List<VarSymbol> parameters = method.getParameters();
+    List<VarSymbol> parameters = ASTHelpers.getSymbol(tree).getParameters();
     int parameterSize = parameters.size() - 1; // -1 means 'formatString' is not parameter
     int formatIndex = 0;
     if (equals(parameters.get(0), "org.slf4j.Marker")) {
       parameterSize--;
       formatIndex = 1;
     }
-    if (equals(parameters.get(parameters.size() - 1), "java.lang.Throwable")) {
+    if (isThrowable(tree.getArguments().get(parameters.size() - 1), state)) {
       parameterSize--;
     }
     if (parameterSize <= 0) {
       return Description.NO_MATCH;
     }
 
-    String format = parameters.get(formatIndex).getConstantValue().toString();
-    if (format == null) {
+    Object constant = ASTHelpers.constValue(tree.getArguments().get(formatIndex));
+    if (constant == null) {
       // format is not resolved at compile-phase
       return Description.NO_MATCH;
     }
+    String format = constant.toString();
 
     int placeholders = countPlaceholder(format);
     if (parameterSize != placeholders) {
@@ -82,7 +86,13 @@ public class PlaceholderMismatch extends BugChecker implements MethodInvocationT
     return Description.NO_MATCH;
   }
 
-  boolean equals(VarSymbol symbol, String name) {
+  private static final com.google.errorprone.matchers.Matcher<ExpressionTree> javaLangThrowable = isSubtypeOf("java.lang.Throwable");
+
+  private boolean isThrowable(ExpressionTree expressionTree, VisitorState state) {
+	return javaLangThrowable.matches(expressionTree, state);
+}
+
+boolean equals(VarSymbol symbol, String name) {
     return name.equals(symbol.getQualifiedName().toString());
   }