+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog follows [Keep a Changelog v1.0.0](https://keepachangelog.com/en/1
 
 ## Unreleased
 
+### Fixed
+
+* Many bugs in PlaceholderMismatch bug pattern
+
 ## v0.1.0 - 2018-07-04
 
 This is the first release, with following bug patterns: